% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_counter.R
\name{counter_init}
\alias{counter_init}
\title{Counter constructor.}
\usage{
counter_init(names = NULL)
}
\arguments{
\item{names}{Character vector of names to add to the new counter.}
}
\value{
A new counter object.
}
\description{
Not a user-side function. Do not invoke directly.
}
\details{
Creates a counter object as described at
\url{https://books.ropensci.org/targets-design/classes.html#counter-class}.
}
\examples{
counter <- counter_init()
counter_set_names(counter, letters)
}
\keyword{internal}
