% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_counter.R
\name{counter_set_names}
\alias{counter_set_names}
\title{Add data to an existing counter object.}
\usage{
counter_set_names(counter, names)
}
\arguments{
\item{counter}{A counter object, defined for internal purposes only.}

\item{names}{Character vector of names to add to the counter.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
counter <- counter_init()
counter_set_names(counter, letters)
}
\keyword{internal}
