% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map_rep_raw.R
\name{tar_append_static_values}
\alias{tar_append_static_values}
\title{Append statically mapped values to target output.}
\usage{
tar_append_static_values(object, values)
}
\arguments{
\item{object}{Return value of a target. Must be a data frame.}

\item{values}{Tibble with the set of static values that the current target
uses.}
}
\description{
For internal use only. Users should not invoke
this function directly.
}
\keyword{internal}
