% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_force.R
\name{tar_force_change}
\alias{tar_force_change}
\title{Convert a condition into a change.}
\usage{
tar_force_change(condition)
}
\description{
Supports \code{\link[=tar_force]{tar_force()}}. This is really an internal function
and not meant to be called by users directly.
}
\keyword{internal}
