% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_group_by.R
\name{tar_group_by_run}
\alias{tar_group_by_run}
\title{Generate a grouped data frame within tar_group_by()}
\usage{
tar_group_by_run(data, by)
}
\arguments{
\item{data}{A data frame to group.}

\item{by}{Nonempty character vector of names of variables to group by.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
