% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_group_count.R
\name{tar_group_count_index}
\alias{tar_group_count_index}
\title{Generate the tar_group column for \code{tar_group_count()}.}
\usage{
tar_group_count_index(data, count)
}
\arguments{
\item{data}{A data frame to group.}

\item{count}{Maximum number of groups.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
