% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_group_count.R
\name{tar_group_count_run}
\alias{tar_group_count_run}
\title{Generate a grouped data frame within tar_group_count()}
\usage{
tar_group_count_run(data, count)
}
\arguments{
\item{data}{A data frame to group.}

\item{count}{Maximum number of rows in each group.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
