% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_group_size.R
\name{tar_group_size_index}
\alias{tar_group_size_index}
\title{Generate the tar_group column for \code{tar_group_size()}.}
\usage{
tar_group_size_index(data, size)
}
\arguments{
\item{data}{A data frame to group.}

\item{size}{Maximum number of rows in each group.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
