% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_group_size.R
\name{tar_group_size_run}
\alias{tar_group_size_run}
\title{Generate a grouped data frame within tar_group_size()}
\usage{
tar_group_size_run(data, size)
}
\arguments{
\item{data}{A data frame to group.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
