% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_knit.R
\name{tar_knit_run}
\alias{tar_knit_run}
\title{Run a \code{knitr} report inside a \code{tar_knit()} target.}
\usage{
tar_knit_run(path, args, deps)
}
\arguments{
\item{path}{Path to the \code{knitr} source file.}

\item{args}{A named list of arguments to \code{knitr::knit()}.}

\item{deps}{An unnamed list of target dependencies of the \code{knitr}
report, automatically created by \code{tar_knit()}.}
}
\value{
Character with the path to the \code{knitr} source file
and the relative path to the output \code{knitr} report.
The output path depends on the input path argument,
which has no default.
}
\description{
Internal function needed for \code{tar_knit()}.
Users should not invoke it directly.
}
\keyword{internal}
