% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map.R
\name{tar_map}
\alias{tar_map}
\title{Static branching.}
\usage{
tar_map(..., values, names)
}
\arguments{
\item{...}{One or more target objects or list of target objects.
Lists can be arbitrarily nested, as in \code{targets::tar_pipeline()}.}

\item{values}{Named list or data frame with values to iterate over.
The names are the names of symbols in the commands and pattern statements,
and the elements are values that get substituted in place of those
symbols. If you want the new values to also be symbols,
use \code{rlang::syms()} on each respective vector or column.}

\item{names}{Character vector, subset of \code{names(values)}
used to generate the suffixes in the names of the new targets.
Can use \code{tidyselect}}
}
\value{
A list of new target objects.
}
\description{
Define collections of new targets.
}
\details{
\code{tar_map()} creates collections of new
targets by iterating over a list of arguments
and substituting symbols into commands and pattern statements.
}
\examples{
\dontrun{
targets::tar_script({
  targets::tar_pipeline(
    tarchetypes::tar_map(
      targets::tar_target(x, a + b),
      targets::tar_target(y, x + a, pattern = map(x))
    )
  )
})
targets::tar_manifest()
}
}
