% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map2_raw.R
\name{tar_map2_group}
\alias{tar_map2_group}
\title{Append the \code{tar_group} variable to a \code{tar_map2()} target.}
\usage{
tar_map2_group(data, group)
}
\arguments{
\item{data}{Data frame to be returned from the target.}

\item{group}{Function on the data to return the \code{tar_group} column.
If \code{group} is \code{NULL}, then no \code{tar_group} column is attached.}
}
\value{
A data frame with a \code{tar_group} column attached (if \code{group}
is not \code{NULL}).
}
\description{
Append the \code{tar_group} variable to a \code{tar_map2()} target.
}
\details{
For internal use only. Users should not invoke
this function directly.
}
\keyword{internal}
