% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map2_raw.R
\name{tar_map2_run}
\alias{tar_map2_run}
\title{Run a dynamic batch of a \code{tar_map2()} target.}
\usage{
tar_map2_run(command, values, columns, parallel_reps)
}
\arguments{
\item{command}{Command to run.}

\item{values}{Data frame of named arguments produced by \code{command1}
that \code{command2} dynamically maps over. Different from the \code{values}
argument of \code{tar_map2()}.}

\item{columns}{tidyselect expression to select columns of \code{values}
to append to the result.}

\item{parallel_reps}{Logical of length 1, whether to run reps within batches
in parallel using the \code{furrr} package. If \code{TRUE}, then \code{furrr} functions
will run the reps in parallel using the \code{future} plan you defined
in your \verb{_targets.R} file. If you are running the pipeline
with \code{tar_make_future()}, then you may need to define a nested plan
with an outer plan to parallelize among targets and an inner
plan to parallelize within targets (e.g. reps).}
}
\value{
A data frame with a \code{tar_group} column attached (if \code{group}
is not \code{NULL}).
}
\description{
Run a dynamic batch of a \code{tar_map2()} target.
}
\details{
For internal use only. Users should not invoke
this function directly.
}
\keyword{internal}
