% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map2_raw.R
\name{tar_map2_run_rep}
\alias{tar_map2_run_rep}
\title{Run a rep in a \code{tar_map2()}-powered function.}
\usage{
tar_map2_run_rep(rep, values, command, batch, seeds, columns, envir)
}
\arguments{
\item{rep}{Rep number.}

\item{values}{Data frame of mapped-over values.}

\item{command}{R command to run.}

\item{batch}{Batch number.}

\item{seeds}{Random number generator seeds of the batch.}

\item{columns}{Expression for appending static columns.}

\item{envir}{Environment of the target.}
}
\value{
The result of running \code{expr}.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
# See the examples of tar_map2_count().
}
\keyword{internal}
