% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_format_nanoparquet.R
\name{tar_nanoparquet_convert}
\alias{tar_nanoparquet_convert}
\title{Nanoparquet convert method}
\usage{
tar_nanoparquet_convert(object, class)
}
\arguments{
\item{object}{R object to convert.}

\item{class}{S3 classes to assign to the returned object.}
}
\description{
Internal function.
}
\keyword{internal}
