% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_format_nanoparquet.R
\name{tar_nanoparquet_read}
\alias{tar_nanoparquet_read}
\title{Nanoparquet read method}
\usage{
tar_nanoparquet_read(path, class)
}
\arguments{
\item{path}{Path to the data.}

\item{class}{S3 classes to assign to the returned object.}
}
\description{
Internal function.
}
\keyword{internal}
