% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_format_nanoparquet.R
\name{tar_nanoparquet_write}
\alias{tar_nanoparquet_write}
\title{Nanoparquet write method}
\usage{
tar_nanoparquet_write(object, path, compression)
}
\arguments{
\item{object}{R object to save.}

\item{path}{Path to the data.}

\item{compression}{Compression type.}
}
\description{
Internal function.
}
\keyword{internal}
