% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_quarto_files.R
\name{tar_quarto_files}
\alias{tar_quarto_files}
\title{Quarto file detection}
\usage{
tar_quarto_files(path = ".")
}
\arguments{
\item{path}{Character of length 1, either the file path
to a Quarto source document or the directory path
to a Quarto project. Defaults to the Quarto project in the
current working directory.}
}
\value{
A named list of important file paths in a Quarto project:
source files to scrape for target dependencies, as well as
other input and output files to track for changes.
}
\description{
Detect the important files in a Quarto project.
}
\examples{
lines <- c(
  "---",
  "title: source file",
  "---",
  "Assume these lines are in report.qmd.",
  "```{r}",
  "1 + 1",
  "```"
)
path <- tempfile(fileext = ".qmd")
writeLines(lines, path)
tar_quarto_files(path)
}
\seealso{
Other Literate programming utilities: 
\code{\link{tar_knitr_deps_expr}()},
\code{\link{tar_knitr_deps}()}
}
\concept{Literate programming utilities}
