% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_quarto_rep_raw.R
\name{tar_quarto_rep_run}
\alias{tar_quarto_rep_run}
\title{Render a batch of parameterized Quarto reports
inside a \code{tar_quarto_rep()} target.}
\usage{
tar_quarto_rep_run(args, execute_params, deps)
}
\arguments{
\item{args}{A named list of arguments to \code{quarto::quarto_render()}.}

\item{execute_params}{A data frame of Quarto parameters to branch over.}

\item{deps}{An unnamed list of target dependencies of the Quarto
report, automatically created by \code{tar_quarto_rep()}.}
}
\value{
Character vector with the path to the Quarto
source file and the rendered output file. Both paths
depend on the input source path, and they have no defaults.
}
\description{
Internal function needed for \code{tar_quarto()}.
Users should not invoke it directly.
}
\keyword{internal}
