% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_quarto_raw.R
\name{tar_quarto_sources}
\alias{tar_quarto_sources}
\title{Detect R source documents in a Quarto project.}
\usage{
tar_quarto_sources(input = NULL)
}
\arguments{
\item{input}{Character vector, \code{input} argument of \code{quarto::quarto_render()}
containing the name of the Quarto source document or the root
directory of the Quarto project.}
}
\value{
A character vector of relative file paths to the Quarto source
documents of the project.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
tar_quarto_sources(".")
}
\keyword{internal}
