% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_rep_raw.R
\name{tar_render_rep_rep}
\alias{tar_render_rep_rep}
\title{Run a rep in a \code{tar_render_rep()}.}
\usage{
tar_render_rep_rep(rep, params, args, path, seeds)
}
\arguments{
\item{rep}{Rep number.}

\item{params}{R Markdown parameters.}

\item{args}{Arguments to \code{rmarkdown::render()}.}

\item{path}{R Markdown output file.}

\item{seeds}{Random number generator seeds of the batch.}
}
\value{
Output file paths.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
# See the examples of tar_quarto_rep().
}
\keyword{internal}
