% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_rep_raw.R
\name{tar_render_rep_run}
\alias{tar_render_rep_run}
\title{Render a batch of parameterized R Markdown reports
inside a \code{tar_render_rep()} target.}
\usage{
tar_render_rep_run(path, params, args, deps, parallel_reps)
}
\arguments{
\item{path}{Path to the R Markdown source file.}

\item{args}{A named list of arguments to \code{rmarkdown::render()}.}

\item{deps}{An unnamed list of target dependencies of the R Markdown
report, automatically created by \code{tar_render_rep()}.}

\item{parallel_reps}{Logical of length 1, whether to run reps within batches
in parallel using the \code{furrr} package. If \code{TRUE}, then \code{furrr} functions
will run the reps in parallel using the \code{future} plan you defined
in your \verb{_targets.R} file. If you are running the pipeline
with \code{tar_make_future()}, then you may need to define a nested plan
with an outer plan to parallelize among targets and an inner
plan to parallelize within targets (e.g. reps).}
}
\value{
Character vector with the path to the R Markdown
source file and the rendered output file. Both paths
depend on the input source path, and they have no defaults.
}
\description{
Internal function needed for \code{tar_render()}.
Users should not invoke it directly.
}
\keyword{internal}
