% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_rep_raw.R
\name{tar_render_rep_run_params}
\alias{tar_render_rep_run_params}
\title{Prepare R Markdown parameters for \code{tar_render_rep()}.}
\usage{
tar_render_rep_run_params(params, batches)
}
\arguments{
\item{params}{Data frame of R Markdown parameters.}

\item{batches}{Number of batches to split up the renderings.}
}
\value{
A batched data frame of R Markdown parameters.
}
\description{
Internal function needed for \code{tar_render_rep()}.
Users should not invoke it directly.
}
\keyword{internal}
