% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render.R
\name{tar_render_run}
\alias{tar_render_run}
\title{Render an R Markdown report inside a \code{tar_render()} target.}
\usage{
tar_render_run(path, args, deps)
}
\arguments{
\item{path}{Path to the R Markdown source file.}

\item{args}{A named list of arguments to \code{rmarkdown::render()}.}

\item{deps}{An unnamed list of target dependencies of the R Markdown
report, automatically created by \code{tar_render()}.}
}
\description{
Internal function needed for \code{tar_render()}.
Users should not invoke it directly.
}
\keyword{internal}
