% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_rep2_raw.R
\name{tar_rep2_run}
\alias{tar_rep2_run}
\title{Run \code{\link[=tar_rep2]{tar_rep2()}} batches.}
\usage{
tar_rep2_run(command, batches, iteration)
}
\arguments{
\item{command}{R expression, the command to run on each rep.}

\item{batches}{Named list of batch data to map over.}

\item{iteration}{Iteration method: \code{"list"}, \code{"vector"}, or \code{"group"}.}
}
\value{
The result of batched replication.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
