% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_rep2_raw.R
\name{tar_rep2_run_rep}
\alias{tar_rep2_run_rep}
\title{Run a rep in a \code{tar_rep2()}-powered function.}
\usage{
tar_rep2_run_rep(rep, slice, command, batch, seeds, envir)
}
\arguments{
\item{rep}{Rep number.}

\item{slice}{Slice of the upstream batch data of the given rep.}

\item{command}{R command to run.}

\item{batch}{Batch number.}

\item{seeds}{Random number generator seeds of the batch.}

\item{envir}{Environment of the target.}
}
\value{
The result of running \code{expr}.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
# See the examples of tar_rep2().
}
\keyword{internal}
