% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_rep_raw.R
\name{tar_rep_run_map_rep}
\alias{tar_rep_run_map_rep}
\title{Run a rep in \code{tar_rep()}.}
\usage{
tar_rep_run_map_rep(rep, expr, batch, reps)
}
\arguments{
\item{rep}{Rep number.}

\item{expr}{R expression to run.}

\item{batch}{Batch number.}

\item{reps}{Number of total reps.}
}
\value{
The result of running \code{expr}.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
# See the examples of tar_rep().
}
\keyword{internal}
