% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_tangle.R
\name{tar_tangle}
\alias{tar_tangle}
\title{Convert Quarto or R Markdown to a pipeline}
\usage{
tar_tangle(path)
}
\arguments{
\item{path}{File path to the literate programming source file.
The file can be a Quarto or R Markdown document.}
}
\value{
A list of new target objects.
See the "Target objects" section for background.
}
\description{
Convert a literate programming source file
into a \code{targets} pipeline.
}
\details{
The word "tangle" comes from the early days of literate
programming (see Knuth 1984).
Tangling is the process of converting a literate
programming source document into machine-readable code.
For example, both \code{knitr::knit(tangle = TRUE)} and \code{\link[knitr:knit]{knitr::purl()}}
accept an \code{.Rmd} file and return an imperative R script with just
the code from the R chunks.

\code{tar_tangle()} is similar, but for a \code{targets} pipeline.
It accepts an R Markdown or Quarto source file as input,
and it returns a list of target definition objects.
Each target definition object comes from evaluating
\code{\link[targets:tar_target]{targets::tar_target()}} on the each of the assignment statements
in each R code chunk in the file. A chunk can look like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ```\{r, deployment = "main"\}
  #| pattern: map(data)
  #| format: qs
  #| cue: tar_cue(mode = "always")
  target_name <- command_to_run(data)
  ```
}\if{html}{\out{</div>}}

And \code{tar_tangle()} will insert this target into the pipeline:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  tar_target(
    name = target_name,
    command = command_to_run(data),
    pattern = map(data),
    format = "qs",
    deployment = "main",
    cue = tar_cue(mode = "always")
  )
}\if{html}{\out{</div>}}

See the example in this help file for a demonstration.
}
\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
targets::tar_dir({  # tar_dir() runs code from a temporary directory.
readr::write_csv(airquality, "data.csv")
lines <- c(
  "---",
  "title: \"Example pipeline\"",
  "---",
  "",
  "```{r}",
  "#| format: file",
  "file <- \"data.csv\"",
  "```",
  "",
  "```{r}",
  "#| memory: persistent",
  "#| packages: [dplyr, readr]",
  "data <- read_csv(file, col_types = cols()) |>",
  "  filter(!is.na(Ozone))",
  "```",
  "",
  "```{r}",
  "#| format: qs",
  "#| cue: tar_cue(mode = \"never\")",
  "model <- lm(Ozone ~ Temp, data) |>",
  "  coefficients()",
  "```",
  "",
  "```{r}",
  "#| deployment: main",
  "#| packages: ggplot2",
  "plot <- ggplot(data) +",
  "  geom_point(aes(x = Temp, y = Ozone)) +",
  "  geom_abline(intercept = model[1], slope = model[2]) +",
  "  theme_gray(24)",
  "```"
)
writeLines(lines, "pipeline.qmd")
targets::tar_script(tarchetypes::tar_tangle("pipeline.qmd"))
targets::tar_make()
targets::tar_read(plot)
})
}
}
\references{
\itemize{
\item Knuth, Donald E. (1984). "Literate Programming".
The Computer Journal. 27 (2). British Computer Society: 97-11.
\url{doi:10.1093/comjnl/27.2.97}.
}
}
\seealso{
Other Domain-specific languages for pipeline construction: 
\code{\link{tar_assign}()},
\code{\link{tar_translate}()}
}
\concept{Domain-specific languages for pipeline construction}
