% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_translate.R
\name{tar_translate}
\alias{tar_translate}
\title{Translate expression to pipeline}
\usage{
tar_translate(code, ...)
}
\arguments{
\item{code}{An expression with ordinary R statements that assign
values to variables.
This code must use assignment operators \verb{<-}, \code{=}, or \verb{->} at the
top level, but it need not call any functions in \code{targets}.
See the example for details.}
}
\value{
A list of \code{tar_target()} objects.
See the "Target objects" section for background.
}
\description{
Translate an ordinary non-\code{targets} imperative expression
into a declarative list of targets for a pipeline.
}
\section{Target objects}{

Most \code{tarchetypes} functions are target factories,
which means they return target objects
or lists of target objects.
Target objects represent skippable steps of the analysis pipeline
as described at \url{https://books.ropensci.org/targets/}.
Please read the walkthrough at
\url{https://books.ropensci.org/targets/walkthrough.html}
to understand the role of target objects in analysis pipelines.

For developers,
\url{https://wlandau.github.io/targetopia/contributing.html#target-factories}
explains target factories (functions like this one which generate targets)
and the design specification at
\url{https://books.ropensci.org/targets-design/}
details the structure and composition of target objects.
}

\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
targets::tar_dir({ # tar_dir() runs code from a temporary directory.
write.csv(airquality, "data.csv", row.names = FALSE)
targets::tar_script({
  library(tarchetypes)
  tar_option_set(packages = c("readr", "dplyr", "ggplot2"))
  tar_translate({
    data <- airquality |>
      filter(!is.na(Ozone)) |>
      tar_target()

    model = lm(Ozone ~ Temp, data) |>
      coefficients() |>
      tar_target()

    plot <- ggplot(data) +
      geom_point(aes(x = Temp, y = Ozone)) +
      geom_abline(intercept = model[1], slope = model[2]) +
      theme_gray(24)
  }, memory = "transient")
})
targets::tar_make()
})
}
}
\seealso{
Other Domain-specific languages for pipeline construction: 
\code{\link{tar_assign}()}
}
\concept{Domain-specific languages for pipeline construction}
