% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tarchetypes-package}
\alias{tarchetypes-package}
\title{targets: Archetypes for Targets}
\description{
The targets package is a pipeline toolkit that brings together
function-oriented programming and Make-like declarative pipelines for
Statistics and data science in R. The tarchetypes packages provides
convenient helper functions to create specialized targets, making
pipelines in targets easier and cleaner to write and understand.
}
