% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_force.R
\name{tar_force_change}
\alias{tar_force_change}
\title{Convert a condition into a change.}
\usage{
tar_force_change(condition)
}
\arguments{
\item{condition}{Logical, whether to run the downstream target
in \code{\link[=tar_force]{tar_force()}}.}
}
\value{
A hash that changes when the downstream target is supposed to run.
}
\description{
Supports \code{\link[=tar_force]{tar_force()}}. This is really an internal function
and not meant to be called by users directly.
}
\keyword{internal}
