% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_map2_raw.R
\name{tar_map2_run}
\alias{tar_map2_run}
\title{Run a dynamic batch of a \code{tar_map2()} target.}
\usage{
tar_map2_run(command, values, columns)
}
\arguments{
\item{command}{Command to run.}

\item{values}{Data frame of named arguments produced by \code{command1}
that \code{command2} dynamically maps over. Different from the \code{values}
argument of \code{tar_map2()}.}

\item{columns}{tidyselect expression to select columns of \code{values}
to append to the result.}
}
\value{
A data frame with a \code{tar_group} column attached (if \code{group}
is not \code{NULL}).
}
\description{
Run a dynamic batch of a \code{tar_map2()} target.
}
\details{
For internal use only. Users should not invoke
this function directly.
}
\keyword{internal}
