% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_rep_raw.R
\name{tar_render_rep_run}
\alias{tar_render_rep_run}
\title{Render a batch of parameterized R Markdown reports
inside a \code{tar_render_rep()} target.}
\usage{
tar_render_rep_run(path, params, args, deps)
}
\arguments{
\item{path}{Path to the R Markdown source file.}

\item{args}{A named list of arguments to \code{rmarkdown::render()}.}

\item{deps}{An unnamed list of target dependencies of the R Markdown
report, automatically created by \code{tar_render_rep()}.}
}
\value{
Character vector with the path to the R Markdown
source file and the rendered output file. Both paths
depend on the input source path, and they have no defaults.
}
\description{
Internal function needed for \code{tar_render()}.
Users should not invoke it directly.
}
\keyword{internal}
