% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_rep.R
\name{tar_rep_run}
\alias{tar_rep_run}
\title{Run a batch in a \code{tar_rep()} archetype.}
\usage{
tar_rep_run(command, batch, reps, iteration)
}
\arguments{
\item{command}{Expression object, command to replicate.}

\item{batch}{Numeric, batch index.}

\item{reps}{Numeric, number of reps per batch.}

\item{iteration}{Character, iteration method.}
}
\value{
Aggregated results of multiple executions of the command.
}
\description{
Internal function needed for \code{tar_rep()}.
Users should not invoke it directly.
}
\keyword{internal}
