% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_package.R
\name{tarchetypes-package}
\alias{tarchetypes-package}
\title{targets: Archetypes for Targets}
\description{
A pipeline toolkit for R, the \code{targets} package brings together
function-oriented programming and Make-like declarative pipelines for
Statistics and data science. The \code{tarchetypes} package provides
convenient helper functions to create specialized targets, making
pipelines in targets easier and cleaner to write and understand.
}
